image_translator/requirements.txt
paddle==2.5.2
paddleocr>=2.6.0.1
opencv-python>=4.7.0
numpy>=1.22.0
requests>=2.28.0
pillow>=9.0.0
python-dotenv>=0.20.0
matplotlib>=3.5.0 